Gestion de la MC - Correction du filtrage des paquets, du parsing XML et de la gestion du débit

+++ Artemis_Core/src/root/elements/network/modules/machine/Machine.java
@@ -23,7 +23,7 @@ public class Machine extends NetworkModule {
 	public Link[] portsInput;
 	
 	/* Time for analysing a packet */
-	public int analyseTime;
+	public double analyseTime;
 	/* Currently transmitted message */
 	public ISchedulable currentlyTransmittedMsg;
